test(mux-player): break up player tests due to playwright memory issues.
