+++ packages/mux-player/test/player.test.js
@@ -1,8 +1,7 @@
-import { fixture, assert, aTimeout, waitUntil, oneEvent, nextFrame } from '@open-wc/testing';
+import { fixture, assert, aTimeout, waitUntil, oneEvent } from '@open-wc/testing';
 import '../src/index.ts';
 
 const isSafari = /.*Version\/.*Safari\/.*/.test(navigator.userAgent);
-const isFirefox = /Firefox/i.test(navigator.userAgent);
 
 // Media Chrome uses a ResizeObserver which ends up throwing in Firefox and Safari in some cases
 // so we want to catch those. It is supposedly not a blocker if this error is thrown.
@@ -851,7 +850,9 @@ describe('<mux-player> playbackId transitions', () => {
     const oldLogError = console.error;
     const oldLogWarn = console.warn;
 
+    // eslint-disable-next-line
     console.error = () => {};
+    // eslint-disable-next-line
     console.warn = () => {};
 
     const player = await fixture(`<mux-player
@@ -882,7 +883,9 @@ describe('<mux-player> playbackId transitions', () => {
     const oldLogError = console.error;
     const oldLogWarn = console.warn;
 
+    // eslint-disable-next-line
     console.error = () => {};
+    // eslint-disable-next-line
     console.warn = () => {};
 
     const player = await fixture(`<mux-player
@@ -997,400 +1000,5 @@ describe('<mux-player> seek to live behaviors', function () {
     const mcPlayEl = playerEl.mediaTheme.shadowRoot.querySelector('media-play-button');
     mcPlayEl.click();
     await waitUntil(() => playerEl.inLiveWindow, 'clicking play did not seek to live window');
-  });
-});
-
-// skip these cue shifting tests except in Firefox
-(isFirefox ? describe : describe.skip)('<mux-player> should move cues up', function () {
-  this.timeout(20000);
-
-  it('when user the user active', async function () {
-    const player = await fixture(`<mux-player
-      playback-id="qP5Eb2cj7MrNnoxBGz012pbZkMHqpIcrKMzd7ykGr01gM"
-      stream-type="on-demand"
-      muted
-      autoplay
-      preload="auto"
-    ></mux-player>`);
-
-    // Wait for us to have at least one showing subtitle/caption
-    await waitUntil(() => Array.prototype.some.call(player.textTracks, (track) => track.mode === 'showing'));
-    // Make sure we're playing
-    await waitUntil(() => !player.paused);
-    // Find the currently showing track
-    const track = Array.prototype.find.call(player.textTracks, (track) => track.mode === 'showing');
-    // Wait for it to have cues added
-    await waitUntil(() => track.cues.length, 2000);
-    const firstCue = track.cues[0];
-    assert.equal(firstCue.line, 'auto', 'the first cue.line is set to auto');
-
-    // Seek to the first Cue so it will be visible/active
-    player.currentTime = firstCue.startTime + 0.1;
-    await waitUntil(() => track.activeCues.length, 2000);
-    const activeCue = track.activeCues[0];
-
-    // Confirm we're inactive initially, as this is a precondition for the test passing.
-    assert(player.mediaController.hasAttribute('userinactive'), 'userinactive is a test precondition');
-    assert.equal(activeCue?.line, 'auto', 'the active cue.line should still be set to auto');
-
-    // Test going from userinactive to user active cue shift
-    player.mediaController.toggleAttribute('userinactive', false);
-    player.dispatchEvent(new Event('userinactivechange'));
-    // Waiting for the condition and then asserting it due to async behavior
-    await waitUntil(() => activeCue.line !== 'auto', 2000);
-    assert.equal(activeCue.line, -4, 'the line is shifted to -4 when user is active');
-
-    // Test going from user active to userinactive active cue unshift
-    player.mediaController.toggleAttribute('userinactive', true);
-    player.dispatchEvent(new Event('userinactivechange'));
-    // Waiting for the condition and then asserting it due to async behavior
-    await waitUntil(() => activeCue.line !== -4, 2000);
-    assert.equal(activeCue.line, 'auto', 'the line is reset to auto when userinactive');
-  });
-
-  it('when the player is paused even if user is inactive', async function () {
-    let done;
-    const promise = new Promise((resolve) => {
-      done = resolve;
-    });
-    const player = await fixture(`<mux-player
-      playback-id="qP5Eb2cj7MrNnoxBGz012pbZkMHqpIcrKMzd7ykGr01gM"
-      stream-type="on-demand"
-      muted
-      preload="auto"
-    ></mux-player>`);
-
-    const mc = player.mediaController;
-    const media = mc.media;
-
-    media.textTracks.addEventListener('addtrack', (e) => {
-      // wait till subtitles have loaded
-      if (e.track.kind === 'subtitles') {
-        // pool until cues have loaded
-        const poolInterval = setInterval(() => {
-          if (e.track.cues?.length) {
-            clearInterval(poolInterval);
-          } else {
-            return;
-          }
-
-          const firstCue = e.track.cues[0];
-          assert.equal(firstCue.line, 'auto', "the first cue's line is set to auto");
-
-          assert.isTrue(player.paused, 'player is paused');
-
-          e.track.addEventListener(
-            'cuechange',
-            () => {
-              const activeCue = e.track.activeCues[0];
-              assert.equal(activeCue.line, -4, "the active cue's line is set to -4");
-              done();
-            },
-            { once: true }
-          );
-
-          media.currentTime = firstCue.startTime + 0.1;
-        }, 10);
-      }
-    });
-
-    return promise;
-  });
-
-  it('unless the cues should be ignored', async function () {
-    let done;
-    const promise = new Promise((resolve) => {
-      done = resolve;
-    });
-    const player = await fixture(`<mux-player
-      playback-id="qP5Eb2cj7MrNnoxBGz012pbZkMHqpIcrKMzd7ykGr01gM"
-      stream-type="on-demand"
-      muted
-      preload="auto"
-    ></mux-player>`);
-
-    const mc = player.mediaController;
-    const media = mc.media;
-
-    media.textTracks.addEventListener('addtrack', (e) => {
-      // wait till subtitles have loaded
-      if (e.track.kind === 'subtitles') {
-        // pool until cues have loaded
-        const poolInterval = setInterval(() => {
-          if (e.track.cues?.length) {
-            clearInterval(poolInterval);
-          } else {
-            return;
-          }
-
-          const firstCue = e.track.cues[0];
-
-          // position first cue at the top of the displayed area
-          // this should currently be ignored
-          firstCue.line = 0;
-
-          assert.isTrue(player.paused, 'player is paused');
-
-          e.track.addEventListener(
-            'cuechange',
-            () => {
-              const activeCue = e.track.activeCues[0];
-              assert.equal(activeCue.line, 0, "the active cue's line was not updated");
-              done();
-            },
-            { once: true }
-          );
-
-          media.currentTime = firstCue.startTime + 0.1;
-        }, 10);
-      }
-    });
-
-    return promise;
-  });
-});
-
-// NOTE: When we fully deprecate "old" stream types ("ll-" prefix and "dvr"), we should update these tests accordingly.
-describe('Feature: stream types & related (including non-media-ui-extension types', async () => {
-  it('should set expected default values for stream type on-demand', async () => {
-    const muxPlayerEl = await fixture(`<mux-player
-      stream-type="on-demand"
-    ></mux-player>`);
-    assert.equal(muxPlayerEl.streamType, 'on-demand');
-    assert(Number.isNaN(muxPlayerEl.targetLiveWindow), 'targetLiveWindow should be NaN for on-demand');
-    assert.equal(muxPlayerEl.media.streamType, 'on-demand');
-  });
-
-  it('should set expected default values for stream type live', async () => {
-    const muxPlayerEl = await fixture(`<mux-player
-      stream-type="live"
-    ></mux-player>`);
-    assert.equal(muxPlayerEl.streamType, 'live');
-    assert(!muxPlayerEl.targetLiveWindow, '!targetLiveWindow for live');
-    assert.equal(muxPlayerEl.media.streamType, 'live');
-  });
-
-  it('should set expected default values for stream type ll-live', async () => {
-    const muxPlayerEl = await fixture(`<mux-player
-      stream-type="ll-live"
-    ></mux-player>`);
-    assert.equal(muxPlayerEl.streamType, 'll-live');
-    assert(!muxPlayerEl.targetLiveWindow, '!targetLiveWindow for ll-live');
-    // Apply the "translated" stream type to underlying `<mux-video>` instance.
-    assert.equal(muxPlayerEl.media.streamType, 'live');
-  });
-
-  it('should set expected default values for stream type live:dvr', async () => {
-    const muxPlayerEl = await fixture(`<mux-player
-      stream-type="live:dvr"
-    ></mux-player>`);
-    assert.equal(muxPlayerEl.streamType, 'live:dvr');
-    assert.equal(muxPlayerEl.targetLiveWindow, Number.POSITIVE_INFINITY);
-    // Apply the "translated" stream type to underlying `<mux-video>` instance.
-    assert.equal(muxPlayerEl.media.streamType, 'live');
-  });
-
-  it('should set expected default values for stream type ll-live:dvr', async () => {
-    const muxPlayerEl = await fixture(`<mux-player
-      stream-type="ll-live:dvr"
-    ></mux-player>`);
-    assert.equal(muxPlayerEl.streamType, 'll-live:dvr');
-    assert.equal(muxPlayerEl.targetLiveWindow, Number.POSITIVE_INFINITY);
-    // Apply the "translated" stream type to underlying `<mux-video>` instance.
-    assert.equal(muxPlayerEl.media.streamType, 'live');
-  });
-
-  it('should apply noautoseektolive to theme for DVR stream types', async () => {
-    const muxPlayerEl = await fixture(`<mux-player
-    ></mux-player>`);
-    assert(!muxPlayerEl.mediaTheme.hasAttribute('noautoseektolive'));
-    muxPlayerEl.streamType = 'live:dvr';
-    assert(muxPlayerEl.mediaTheme.hasAttribute('noautoseektolive'));
-    muxPlayerEl.streamType = undefined;
-    assert(!muxPlayerEl.mediaTheme.hasAttribute('noautoseektolive'));
-    muxPlayerEl.streamType = 'll-live:dvr';
-    assert(muxPlayerEl.mediaTheme.hasAttribute('noautoseektolive'));
-  });
-});
-
-describe('Feature: cuePoints', async () => {
-  it('adds cuepoints', async () => {
-    const cuePoints = [
-      { time: 0, value: { label: 'CTA 1', showDuration: 10 } },
-      { time: 15, value: { label: 'CTA 2', showDuration: 5 } },
-      { time: 21, value: { label: 'CTA 3', showDuration: 2 } },
-    ];
-    const playbackId = '23s11nz72DsoN657h4314PjKKjsF2JG33eBQQt6B95I';
-    const muxPlayerEl = await fixture(`<mux-player
-      stream-type="on-demand"
-      playback-id="${playbackId}"
-    ></mux-player>`);
-    if (muxPlayerEl.readyState === 0) {
-      await oneEvent(muxPlayerEl, 'loadedmetadata');
-    }
-    await muxPlayerEl.addCuePoints(cuePoints);
-    assert.deepEqual(muxPlayerEl.cuePoints, cuePoints);
-  });
-
-  it('dispatches a cuepointchange event when the active cuepoint changes', async function () {
-    this.timeout(10000);
-
-    const cuePoints = [
-      { time: 0, value: { label: 'CTA 1', showDuration: 10 } },
-      { time: 15, value: { label: 'CTA 2', showDuration: 5 } },
-      { time: 21, value: { label: 'CTA 3', showDuration: 2 } },
-    ];
-    const playbackId = '23s11nz72DsoN657h4314PjKKjsF2JG33eBQQt6B95I';
-    const muxPlayerEl = await fixture(`<mux-player
-      stream-type="on-demand"
-      playback-id="${playbackId}"
-    ></mux-player>`);
-    if (muxPlayerEl.readyState === 0) {
-      await oneEvent(muxPlayerEl, 'loadedmetadata');
-    }
-    await muxPlayerEl.addCuePoints(cuePoints);
-    const expectedCuePoint = cuePoints[1];
-    muxPlayerEl.currentTime = expectedCuePoint.time + 0.1;
-    let event = await oneEvent(muxPlayerEl, 'cuepointchange');
-    // NOTE: For Safari/WebKit, we may actually get an event for
-    // the first cue before getting one for the second after the seek.
-    // This is valid and should be accounted for (CJP).
-    if (event.detail?.time === cuePoints[0].time) {
-      event = await oneEvent(muxPlayerEl, 'cuepointchange');
-    }
-    assert.equal(event.target, muxPlayerEl, 'event target should be the MuxPlayerElement instance');
-    assert.deepEqual(event.detail, expectedCuePoint);
-    assert.deepEqual(muxPlayerEl.activeCuePoint, expectedCuePoint);
-  });
-
-  it('clears cuepoints when playback-id is updated', async () => {
-    const cuePoints = [
-      { time: 0, value: { label: 'CTA 1', showDuration: 10 } },
-      { time: 15, value: { label: 'CTA 2', showDuration: 5 } },
-      { time: 21, value: { label: 'CTA 3', showDuration: 2 } },
-    ];
-    const playbackId = '23s11nz72DsoN657h4314PjKKjsF2JG33eBQQt6B95I';
-    const muxPlayerEl = await fixture(`<mux-player
-      stream-type="on-demand"
-      playback-id="${playbackId}"
-    ></mux-player>`);
-    if (muxPlayerEl.readyState === 0) {
-      await oneEvent(muxPlayerEl, 'loadedmetadata');
-    }
-    await muxPlayerEl.addCuePoints(cuePoints);
-    assert.deepEqual(muxPlayerEl.cuePoints, cuePoints, 'cue points were added as expected');
-    muxPlayerEl.playbackId = 'DS00Spx1CV902MCtPj5WknGlR102V5HFkDe';
-    await oneEvent(muxPlayerEl, 'emptied');
-    // Safari needs an extra tick for the cues to clear
-    await nextFrame();
-    assert.equal(muxPlayerEl.cuePoints.length, 0, 'cuePoints should be empty');
-  });
-});
-
-describe('Feature: chapters', async () => {
-  it('adds chapters', async () => {
-    const chapters = [
-      { startTime: 0, endTime: 5, value: 'Chapter 1' },
-      { startTime: 5, endTime: 10, value: 'Chapter 2' },
-      { startTime: 10, endTime: 15, value: 'Chapter 3' },
-    ];
-    const playbackId = '23s11nz72DsoN657h4314PjKKjsF2JG33eBQQt6B95I';
-    const muxPlayerEl = await fixture(`<mux-player
-      stream-type="on-demand"
-      playback-id="${playbackId}"
-    ></mux-player>`);
-
-    await oneEvent(muxPlayerEl, 'loadedmetadata');
-    await muxPlayerEl.addChapters(chapters);
-
-    // need a timeout for Safari/webkit
-    await nextFrame();
-
-    assert.deepEqual(muxPlayerEl.chapters, chapters);
-  });
-
-  it('dispatches a chapterchange event when the active chapter changes', async function () {
-    this.timeout(10000);
-    const chapters = [
-      { startTime: 0, endTime: 5, value: 'Chapter 1' },
-      { startTime: 5, endTime: 10, value: 'Chapter 2' },
-      { startTime: 10, endTime: 15, value: 'Chapter 3' },
-    ];
-    const playbackId = '23s11nz72DsoN657h4314PjKKjsF2JG33eBQQt6B95I';
-    const muxPlayerEl = await fixture(`<mux-player
-      stream-type="on-demand"
-      playback-id="${playbackId}"
-    ></mux-player>`);
-    await oneEvent(muxPlayerEl, 'loadedmetadata');
-    await muxPlayerEl.addChapters(chapters);
-    const expectedChapter = chapters[1];
-    muxPlayerEl.currentTime = expectedChapter.startTime + 0.1;
-    let event = await oneEvent(muxPlayerEl, 'chapterchange');
-    // NOTE: For Safari/WebKit, we may actually get an event for
-    // the first cue before getting one for the second after the seek.
-    // This is valid and should be accounted for (CJP).
-    if (event.detail?.startTime === chapters[0].startTime) {
-      event = await oneEvent(muxPlayerEl, 'chapterchange');
-    }
-    assert.equal(event.target, muxPlayerEl, 'event target should be the MuxPlayerElement instance');
-    assert.deepEqual(event.detail, expectedChapter);
-    assert.deepEqual(muxPlayerEl.activeChapter, expectedChapter);
-  });
-
-  it('clears chapters when playback-id is updated', async function () {
-    this.timeout(5000);
-    const chapters = [
-      { startTime: 0, endTime: 5, value: 'Chapter 1' },
-      { startTime: 5, endTime: 10, value: 'Chapter 2' },
-      { startTime: 10, endTime: 15, value: 'Chapter 3' },
-    ];
-    const playbackId = '23s11nz72DsoN657h4314PjKKjsF2JG33eBQQt6B95I';
-    const muxPlayerEl = await fixture(`<mux-player
-      stream-type="on-demand"
-      playback-id="${playbackId}"
-    ></mux-player>`);
-    await oneEvent(muxPlayerEl, 'loadedmetadata');
-    await muxPlayerEl.addChapters(chapters);
-    muxPlayerEl.currentTime = chapters[1].startTime + 0.1;
-    await oneEvent(muxPlayerEl, 'chapterchange');
-    assert.deepEqual(muxPlayerEl.chapters, chapters, 'chapters were added');
-    muxPlayerEl.playbackId = 'DS00Spx1CV902MCtPj5WknGlR102V5HFkDe';
-    await oneEvent(muxPlayerEl, 'emptied');
-    await nextFrame();
-    assert.equal(muxPlayerEl.chapters.length, 0, 'chapters are empty');
-  });
-});
-
-describe('currentPdt and getStartDate', async () => {
-  it('currentPdt and getStartDate work as expected', async function () {
-    this.timeout(5000);
-
-    const player = await fixture(`<mux-player
-      env-key="ilc02s65tkrc2mk69b7q2qdkf"
-      stream-type="on-demand"
-      prefer-playback="mse"
-      muted
-      title="A title"
-      preload="auto"
-    ></mux-player>`);
-
-    player.addEventListener('loadstart', async function () {
-      player.currentTime = 60;
-
-      await nextFrame();
-
-      const currentPdt = player.currentPdt;
-      const startDate = player.getStartDate();
-
-      assert.equal(
-        startDate.getTime(),
-        currentPdt.getTime() - player.currentTime * 1000,
-        'currentPdt should be 60 seconds greater than getStartDate'
-      );
-    });
-
-    await nextFrame();
-
-    player.playbackId = 'UgKrPYAnjMjP6oMF4Kcs1gWVhtgYDR02EHQGnj022X1Xo';
   });
 });
